pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-neo4j>=0.5.0",
    "langchain-openai>=0.3.30",
    "neo4j>=5.28.2",
    "openai>=1.99.9",
    "psycopg2-binary>=2.9.10",
    "rdflib>=7.1.4",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.35.0",
    "watchfiles>=1.1.0",
]
